Converts Eclipse projects into Maven projects
This commit converts the Eclipse projects into Maven projects.
This means that we are not longer on the Eclipse builder to
build our projects, but instead use Maven for their building.
This enables us in a later step to employ automatic dependency
resolution. It also makes the Eclipse build process more similar
to the build process on the command line. Furthermore, we are
fixing some warnings in the pom.xml files with this commit.

Refs #10