Change bundleId, modify plugin.xml and adapt test case
Changes the bundle id to nl.tudelft.watchdog. nl.tudelft.* is whitelisted by
testing tool Q7 to be open-source, so we can use our free organization account
with unlimited test executions to run WatchDog.

Furthermore, the WatchDog start command was malformed in plugin.xml, preventing
an initilization of the WatchDog plugin in the IDE.

Adapts a failing unit test.

Closes #16