Fixes problem with double-transfer of interval
Fixes the problem that the last saved interval is being transfered once on
shutdown of the plugin, and once on restart, because the preference store
was not saved properly. The preference store is now saved on every action
where it could be modified, plus on plugin shutdown. This has potentially
beneficial effects on other (hidden) bugs with the preference mechanism.

Refs #40