Fixes dependency issues
This commit puts the dependency on JFreeChart back into the IDE-specific projects. This is necessary because this library internally depends on SWT an therefore on org.eclipse.swt. However, the core project should not depend on IDE-specific code. Therefore, to remove this dependency and fix the maven build again, all code depending on JFreeChart is put into the IDE-specific project. Unfortunately this results in two identical EventStatistics classes.