Update plugin ID specified for IntelliJ
This makes sure the update of WatchDog will be correctly
pushed to existing users. Originally the plugin ID was
equivalent to the artifact ID in the pom.xml (which is
changed now), but that is not a requirement.
Therefore we can safely change the value in plugin.xml
with a different maven pom.xml ID.